move valid_hostname func into website module
