+++ httpobs/conf/httpobs.conf
@@ -27,7 +27,6 @@ user_agent = Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:98.0) Gecko/201001
 abort_scan_time = 1800
 allow_kickstart = no
 allow_kickstart_num_aborted = 5
-allow_localhost = no
 broker = redis://localhost:6379/0
 broker_reconnection_sleep_time = 15
 cycle_sleep_time = .5
